+++ apps/user/src/app/(bottomNav)/ui/filterTab/type.ts
@@ -1,0 +1,13 @@
+import type { foodTypeList } from "./constant";
+
+export type DayType = "오전" | "오후";
+export type HourType = number & { __brand: "Hour" };
+export type MinuteType = number & { __brand: "Minute" };
+
+export type PickupTime = {
+  day: DayType;
+  hour: HourType;
+  minute: MinuteType;
+};
+
+export type FoodType = (typeof foodTypeList)[number];
